requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
flask==3.0.0
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1
aiohttp==3.9.1
gunicorn==21.2.0
